requirements.txt
torch>=2.0
numpy
h5py
trimesh
shapely
osmnx
requests
pyyaml
matplotlib
seaborn
scikit-learn
plotly
nbformat>=4.2.0
imageio
imageio-ffmpeg
scikit-image
tqdm
scipy
rtree
